GenreList вроде работает
