+++ src/reducers/reducer.js
@@ -1,6 +1,15 @@
 import data from './data';
-import visibilityFilter from './visibilityFilter';
+// import visibilityFilter from './visibilityFilter';
 import {combineReducers} from "redux";
+
+const visibilityFilter = (state = `All genres`, action) => {
+  switch (action.type) {
+    case `SET_VISIBILITY_FILTER`:
+      return action.filter;
+    default:
+      return state;
+  }
+};
 
 export default combineReducers({
   data,
