+++ src/reducers/visibilityFilter.js
@@ -1,10 +1,10 @@
-const visibilityFilter = (state = `All genres`, action) => {
-  switch (action.type) {
-    case `SET_VISIBILITY_FILTER`:
-      return action.filter;
-    default:
-      return state;
-  }
-};
+// const visibilityFilter = (state = `All genres`, action) => {
+//   switch (action.type) {
+//     case `SET_VISIBILITY_FILTER`:
+//       return action.filter;
+//     default:
+//       return state;
+//   }
+// };
 
-export default visibilityFilter;
+// export default visibilityFilter;
